requirements.txt
selenium
requests
python-dotenv
gspread
google-auth
groq
